+++ src/main/java/com/bettopia/game/controller/ChatController.java
@@ -1,48 +1,16 @@
 package com.bettopia.game.controller;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.ui.Model;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
-import com.bettopia.game.model.chatbot.ChatQADTO;
-import com.bettopia.game.model.chatbot.ChatQAService;
-
-@RestController
-@RequestMapping("/api/chat")
+@Controller
+@RequestMapping("/chatbot")
 public class ChatController {
-
-	@Autowired
-	ChatQAService chatService;
-	
-	/*
-	 * @GetMapping(value = "/chat", produces = "text/plain;charset=UTF-8") public
-	 * String f1() { return "test"; }
-	 */
-	
-	// 질문 전체 리스트 반환
-    @GetMapping("/question")
-    public List<ChatQADTO> getAllQuestions() {
-        return chatService.selectAll();
+	@GetMapping("")
+    public String moveToChatView() {
+		System.out.println("진입성공");
+        return "chatbot/chatView";
     }
 
-    // uid로 질문-답변 1개 조회
-    @GetMapping(value= "/answer", produces = "text/plain;charset=UTF-8")
-    public String getAnswerByUid(String uid) {
-        return chatService.questiontByUid(uid);
-    }
-
-    @GetMapping(value="/questionByCate", produces = "application/json;charset=UTF-8")
-    public List<ChatQADTO> getQuestionByCate(@RequestParam("category") String category){
-    	System.out.println("💡 전달받은 카테고리: [" + category + "]");
-        List<ChatQADTO> list = chatService.selectByCate(category.trim());
-        System.out.println("💬 결과 개수: " + list.size());
-
-    	return list;
-    }
-	
 }
