--- src/main/java/com/bettopia/game/controller/ChatController_Test.java
@@ -1,16 +1,0 @@
-package com.bettopia.game.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/chatbot")
-public class ChatController_Test {
-	@GetMapping("")
-    public String moveToChatView() {
-		System.out.println("진입성공");
-        return "chatbot/chatView";
-    }
-
-}
